% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_connect.R
\name{create_websocket}
\alias{create_websocket}
\title{Wrapper around \code{websocket::WebSocket$new}}
\usage{
create_websocket(socket_url, autoConnect = FALSE)
}
\arguments{
\item{socket_url}{Typically always "streamer-ws.tdameritrade.com", this is nevertheless returned
by a successful call to \code{td_get_user_principals()}, nested at \code{user_prins[["streamerInfo"]][["streamerSocketUrl"]]}}

\item{autoConnect}{Default is FALSE, given that we'll typically set up out handler functions (eg
\code{ws$onMessage(), ws$onClose}) before invoking \code{ws$connect()}}
}
\value{
An R6 websocket object, against which methods are called.  See \href{https://github.com/rstudio/websocket}{the rstudio/websocket github repo}
for more info.
}
\description{
Wrapper around \code{websocket::WebSocket$new}
}
