% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_connect.R
\name{credentials_request_string}
\alias{credentials_request_string}
\title{Create json string for inital websocket handshake}
\usage{
credentials_request_string(user_prins, pretty = FALSE)
}
\arguments{
\item{user_prins}{The list of login credentials returned by a successful call to \code{td_get_user_principals()}.}

\item{pretty}{}
}
\value{
json
}
\description{
Create json string for inital websocket handshake
}
