% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datetime_to_ms}
\alias{datetime_to_ms}
\title{Convert a datetime object to milliseconds since unix epoch}
\usage{
datetime_to_ms(dttm)
}
\arguments{
\item{dttm}{}
}
\value{
a double
}
\description{
Convert a datetime object to milliseconds since unix epoch
}
