% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{master_parser}
\alias{master_parser}
\title{Parse one line of received json into a tibble}
\usage{
master_parser(my_json)
}
\arguments{
\item{my_json}{Some json, as received from the TD streaming API}
}
\value{
A tibble
}
\description{
The TD streamer sends json which \code{master_parser()} 1), converts to a nested
list/df object per jsonlite::fromJSON() and then 2) using one of the associated helper functions
\code{parser_*} to flatten into a nicely rectangular tibble
}
