% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ms_to_datetime}
\alias{ms_to_datetime}
\title{Convert milliseconds since unix epoch to a datetime object}
\usage{
ms_to_datetime(milliseconds_since_epoch, tz = "America/Chicago")
}
\arguments{
\item{tz}{}
}
\value{
a Date object
}
\description{
Convert milliseconds since unix epoch to a datetime object
}
\examples{
ms <- 1559235837234
ms_to_datetime(ms)
}
