% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{requestor}
\alias{requestor}
\title{Send requests to TD streamer}
\usage{
requestor(parameter_keys, requestid = "2", service = "QUOTE",
  command = "SUBS", parameter_fields = c(0:11), pretty = FALSE,
  user_prins)
}
\arguments{
\item{parameter_keys}{Vector of symbols requested}

\item{requestid}{Default is "2"}

\item{service}{Defaults to "QUOTE"; Common values are "ADMIN", "QUOTE", "NEWS_HEADLINE", though a
complete listing can be found in section 4.4 of the
\href{https://developer.tdameritrade.com/content/streaming-data#_Toc504640602}{TD Ameritrade streaming API} docs.}

\item{command}{Default is "SUBS"}

\item{parameter_fields}{A numeric vector of requested fields, default is 0:11}

\item{pretty}{Defaults to FALSE. Used for inspecting the output/ debugging.}

\item{user_prins}{some user_principals; a list}
}
\value{
Output will be a string of json that is expected by the streamer
}
\description{
Send requests to TD streamer
}
\examples{
requestor(c("TSLA", "MSFT"))
requestor("TSLA_122118P400", service = "OPTION")
}
