% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_trading.R
\name{td_get_account_balances}
\alias{td_get_account_balances}
\title{Takes a snapshot of account balance}
\usage{
td_get_account_balances(account_no = "489837238", access_token)
}
\arguments{
\item{account_no}{Funciton currently defaults to sidePocket's act #}

\item{access_token}{Valid access_token, typically generated via update_refresh_tokens()}
}
\value{
a tibble()
}
\description{
Takes a snapshot of account balance
}
