% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{td_get_option_chain}
\alias{td_get_option_chain}
\title{Get a term structure of extant option contract symbols}
\usage{
td_get_option_chain(
  symbol = "TSLA",
  apikey = "moonriver@AMER.OAUTHAP",
  contractType = "ALL",
  strikeCount = 30,
  includeQuotes = FALSE,
  strategy = "SINGLE",
  interval = NULL,
  strike = NULL,
  range = "NTM",
  fromDate = Sys.Date(),
  monthsToGet = 12,
  volatility = NULL,
  underlyingPrice = NULL,
  interestRate = NULL,
  daysToExpiration = NULL,
  expMonth = "ALL",
  optionType = "ALL",
  access_token = .token_set$access_token
)
}
\arguments{
\item{symbol}{Symbol of the underlying, eg "TSLA"}

\item{apikey}{The TD developer api key, aka the app name}

\item{contractType}{Default is "ALL"}

\item{strikeCount}{Number of strikes to return.  Default is 30.}

\item{includeQuotes}{Default is FALSE}

\item{strategy}{Default is "SINGLE"}

\item{interval}{Default is NULL}

\item{strike}{Default is NULL}

\item{range}{Default is "NTM" or near-the-money}

\item{monthsToGet}{Number of months into the future to get, defualt is 12}

\item{volatility}{Used for getting options greeks. Default is NULL}

\item{underlyingPrice}{Used for getting options greeks. Default is NULL}

\item{interestRate}{Used for getting options greeks. Default is NULL}

\item{daysToExpiration}{Used for getting options greeks. Default is NULL}

\item{expMonth}{Default is "ALL"}

\item{optionType}{"CALL", "PUT", or "ALL".  Default is "ALL"}

\item{access_token}{A valid access token (they expire within 30 mins of refresh)}
}
\value{
A chr vector of option symbol names.
}
\description{
Get a term structure of extant option contract symbols
}
