% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_trading.R
\name{td_get_orders}
\alias{td_get_orders}
\title{Retrieve outstanding orders}
\usage{
td_get_orders(access_token)
}
\value{
A tibble
}
\description{
Retrieve outstanding orders
}
\examples{
td_get_orders(refresh_tokens$access_tokens)
}
