% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{td_get_price_history}
\alias{td_get_price_history}
\title{Get HLCO price history for a symbol}
\usage{
td_get_price_history(
  symbol = "TSLA",
  apikey = "moonriver@AMER.OAUTHAP",
  periodType = "day",
  period = 1,
  frequencyType = "minute",
  frequency = 1,
  endDate = "",
  startDate = "",
  needExtendedHoursData = "true",
  access_token
)
}
\arguments{
\item{apikey}{The TD developer api key, aka the app name}

\item{periodType}{Valid values are `day`, `month`, `year,` or `ytd`; default is day if blank}

\item{period}{The number of periods to show of PeriodType.}

\item{frequencyType}{Valid frequencyTypes for PeriodType: day: 1,2,3,4,5,10*
month: 1*, 2, 3, 6
year: 1*, 2, 3, 5, 10, 15, 20
ytd: 1*}

\item{frequency}{Valid frequencies by frequencyType: minute: 1*, 5, 10, 15, 30
daily: 1*
weekly: 1*
monthly: 1*}

\item{endDate}{End date as milliseconds since epoch.
If startDate and endDate are provided, period should not be provided.
Default is previous trading day.}

\item{startDate}{Start date as milliseconds since epoch.
If startDate and endDate are provided, period should not be provided.}

\item{needExtendedHoursData}{Default is `true`. `false` for reg mkt hours only}

\item{access_token}{A valid access token (they expire within 30 mins of refresh)}
}
\value{
A tibble of dates and HLCO prices
}
\description{
Get HLCO price history for a symbol
}
\examples{
td_get_price_history("$SPX.X")
}
