% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{td_get_quote}
\alias{td_get_quote}
\title{Get quote for individal stock}
\usage{
td_get_quote(
  symbol = "TSLA",
  apikey = "moonriver@AMER.OAUTHAP",
  access_token = ""
)
}
\arguments{
\item{symbol}{A valid stock ticker}

\item{apikey}{The TD developer api key, aka the app name}

\item{access_token}{A valid access token (they expire within 30 mins of refresh)}
}
\value{
A tibble of prices
}
\description{
Get quote for individal stock
}
