% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{td_get_quotes}
\alias{td_get_quotes}
\title{Get quote for individal stock}
\usage{
td_get_quotes(
  symbols = c("AAPL", "TSLA"),
  apikey = "moonriver@AMER.OAUTHAP",
  access_token
)
}
\arguments{
\item{apikey}{The TD developer api key, aka the app name}

\item{access_token}{A valid access token (they expire within 30 mins of refresh)}

\item{symbol}{A vector of stock tickers}
}
\value{
A tibble of prices
}
\description{
Get quote for individal stock
}
