% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_connect.R
\name{td_get_user_principals}
\alias{td_get_user_principals}
\title{Get user principals method from TD API.}
\usage{
td_get_user_principals(
  access_token,
  fields = "streamerSubscriptionKeys,streamerConnectionInfo"
)
}
\arguments{
\item{access_token}{Authentication token obtained from successful call to \code{td_post_access_token()}}

\item{fields}{Typically always "streamerSubscriptionKeys,streamerConnectionInfo"; See the
\href{https://developer.tdameritrade.com/user-principal/apis/get/userprincipals-0}{TD documentation} for more.}
}
\value{
Returns a nested list of authentication parameters
}
\description{
This is part of the TD authentication handshake process, retrieving authentication
credentials to the streaming TD API.
}
