% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_trading.R
\name{td_order}
\alias{td_order}
\title{Primary trading function USE WITH CAUTION}
\usage{
td_order(
  symbol,
  orderType = "LIMIT",
  quantity = 0,
  price,
  instruction = "BUY",
  account_no = "489837238",
  access_token
)
}
\arguments{
\item{symbol}{A security symbol, eg "IBM", "GS", "TSLA_110119C315"}

\item{orderType}{Can be "LIMIT" or "MARKET}

\item{quantity}{Number of shares OR contracts in case of options}

\item{price}{Applies to LIMIT orders only; the price at which to execute}

\item{instruction}{"BUY" or "SELL" or *MANY* others}

\item{account_no}{TD account #}
}
\value{
The reply to the POST command includes an order # in the header
}
\description{
Primary trading function USE WITH CAUTION
}
\examples{
td_order(symbol = "TSLA", "LIMIT", 10, price = 300, "SELL")
}
