% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_trading.R
\name{td_order_status}
\alias{td_order_status}
\title{This will eventually get order status}
\usage{
td_order_status(access_token)
}
\value{
Will return a tibble
}
\description{
This will eventually get order status
}
