% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_connect.R
\name{td_post_access_token}
\alias{td_post_access_token}
\title{Obtain access or refresh tokens from the TD API}
\usage{
td_post_access_token(grant_type, refresh_token, access_type, code,
  client_id, redirect_uri)
}
\arguments{
\item{grant_type}{The grant type of the oAuth scheme.
Possible values are \code{"authorization_code"} or \code{"refresh_token"}}

\item{refresh_token}{Required if using refresh token grant}

\item{access_type}{Set to \code{"offline"} to receive a refresh token}

\item{code}{Required if trying to use authorization code grant}

\item{client_id}{OAuth User ID of the application, a parameter also known as 'apikey' in the
TD API documentation}

\item{redirect_uri}{Required if trying to use authorization code grant, will typically be
\code{"http://localhost:8080"} on the local machine}
}
\value{
A five element named list: \code{access_token, refresh_token, expires_in, refresh_token_expires_in, token_type }
}
\description{
Obtain access or refresh tokens from the TD API
}
