% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_connect.R
\name{update_refresh_tokens}
\alias{update_refresh_tokens}
\title{Update a token set}
\usage{
update_refresh_tokens(path_to_file = "")
}
\arguments{
\item{path_to_file}{Local file path to token set.}
}
\value{
A new set of tokens.
}
\description{
This function does two things simultaneously: 1) returns an updated set of tokens
and 2), as a side effect, saves/ replaces the old token set via \code{saveRDS()} at the same location
indicated by \code{path_to_file}
}
