% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{url_encode}
\alias{url_encode}
\title{Encode a URL}
\usage{
url_encode(string)
}
\arguments{
\item{string}{A character string}
}
\value{
A character-encoded string, as required by some web APIs.
}
\description{
Encode a URL
}
\examples{
url_encode("http://localhost:8080")
}
