% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{option_name_parser}
\alias{option_name_parser}
\title{This decodes an option contract symbol into its useful constituent data}
\usage{
option_name_parser(option_string)
}
\arguments{
\item{option_string}{A}
}
\value{
A 1x7 tibble with the following fields: stock, strike, type,
expire_month, expire_day. expire_year, and expire_date
}
\description{
This decodes an option contract symbol into its useful constituent data
}
\examples{
option_name_parser("TSLA_110119C315")
}
