% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{option_name_parser}
\alias{option_name_parser}
\title{Decode an option contract symbol into its constituents}
\usage{
option_name_parser(df, key)
}
\arguments{
\item{df}{A dataframe}

\item{key}{The column dame of the df with the option symbols}
}
\value{
Modifies in a place a tibble
}
\description{
The TD API uses a certain format for its options symbology (), and this function
decomposes the symbol into a tibble with seven fields
}
\examples{
option_name_parser(data.frame(key = "TSLA_110119C315"), "key")
}
