% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parser_response, parser_notify, parser_data}
\alias{parser_response, parser_notify, parser_data}
\alias{parser_response}
\alias{parser_response,}
\alias{parser_notify,}
\alias{parser_data}
\alias{parser_notify}
\title{Private helper functions used by master_parser() for}
\usage{
parser_response(raw_parse)

parser_notify(raw_parse)

parser_data(raw_parse)
}
\arguments{
\item{raw_parse}{A data frame returned by jsonlite::fromJSON()}
}
\value{
a data.frame object
}
\description{
The function master_parser() is the handler for a stream of json from the TD API.
Once having been initially parsed by
}
\examples{
parser_response()
}
