% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parser_response}
\alias{parser_response}
\alias{parser_notify}
\alias{parser_data}
\title{Helper functions used within master_parser()}
\usage{
parser_response(raw_parse)

parser_notify(raw_parse)

parser_data(raw_parse)
}
\arguments{
\item{raw_parse}{The heavily nested list/df object returned from \code{jsonlite::fromJSON()}}
}
\value{
A tibble
}
\description{
The function \code{master_parser()} is the handler for a stream of json from the TD API.
Once having been initially parsed by an internal call to \code{jsonlite::fromJSON()}, master_parser
uses a call to \code{switch()} to one of these functions.
}
\examples{
parser_response(jsonlite::fromJSON(some_json))
}
