% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_trading.R
\name{td_delete_order}
\alias{td_delete_order}
\title{This will eventually delete an order}
\usage{
td_delete_order(order_number)
}
\arguments{
\item{order_url}{}
}
\value{
Will return a tibble with a confirm #(?) and a timestamp
}
\description{
This will eventually delete an order
}
\examples{
td_delete_order(order_number = "123456")
}
