% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{td_market_value_traded}
\alias{td_market_value_traded}
\title{Return a df of volume, mkt cap and mkt value traded}
\usage{
td_market_value_traded(symbol)
}
\arguments{
\item{symbol}{}
}
\value{
A tibble of the last n days' trading volume, HLC, and mkt cap
}
\description{
Return a df of volume, mkt cap and mkt value traded
}
\examples{
td_market_value_traded("MSFT")
}
